Initial revision of meta-hosted-mender-login layer
This layer provides a graphical login to Hosted Mender, written in
PyQt5 and thus depends on meta-qt5, that retrieves and stores the
Hosted Mender tenant token used to interface with Mender API. The
application is run each time the unit boots.

I've also included some recipes for configuring ethernet as well as
some (conditional) fixes for the Toradex' 7" touch display, the latter
recipe does not apply unless "touchscreen" added to MACHINE_FEATURES,
and the MACHINE is colibri-imx7.

Signed-off-by: Alf-Rune <alf.rune@northern.tech>